requirements.txt
langchain_openai
langchain
langchain_core
streamlit
python_dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
# groq
# cassio
# beautifulsoup4
# langchain-groq
wikipedia
arxiv
langchain_hub